Major Changes
Navbar look redesigned. Changed text color to black using my the index.css file. Gave opacity to the navbar to make it look transparent.

Used .border-radius mixin to round the corners of the images.

Added top and bottom border to the middle section using the index.css file.

Added a rounded border to the footer.